Add files via upload
This repository contains a TensorFlow-based model for cancer detection, designed for binary classification tasks. The model features an input layer for 30 features, two hidden layers with 32 neurons each utilizing ReLU activation, and an output layer with a sigmoid activation function. It is compiled with the Adam optimizer and trained using binary crossentropy as the loss function. The model is configured for training with 100 epochs and a batch size of 1. This project aims to assist in distinguishing between cancerous and non-cancerous samples, making it a valuable tool for medical diagnosis and research.